Update count.js
not sure if this intended on being used on heroku. I assume after seeing it on there. aftering pulling the most recent update today without doing both of the things i changed in this PR the bot wouldnt boot and count command wouldnt work. didnt have an issue before updating but im not sure if you changed the count command recent or if thats how it has been since ive never edited it before..

Sorry if neither of these are right. I am pretty new and this project has been vary fun to explore with!